% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco_tsp.R
\name{aco_tsp}
\alias{aco_tsp}
\title{ACO Algorithm applied to the Traveling Salesman Problem}
\usage{
aco_tsp(
  x,
  y,
  alpha,
  beta,
  evaporation,
  randomness_factor,
  numb_ants,
  iterations
)
}
\arguments{
\item{x}{X-Values from the Cities}

\item{y}{y-Values from the Cities}

\item{alpha}{The strength of the pheromone. The higher the stronger.}

\item{beta}{Relevance of the distance. The higher the more decreases the relevance.}

\item{evaporation}{Strength with which the pheromones evaporate.}

\item{randomness_factor}{Factor for random iterations.}

\item{numb_ants}{Number of ants.}

\item{iterations}{Number of iterations.}
}
\value{
All the routes with the minimal distance, possible that this are more than one.
}
\description{
ACO Algorithm applied to the Traveling Salesman Problem
}
\author{
The code is adopted from: https://github.com/ciessielski/ACOTSP
}
