% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco.R
\name{c_sigma}
\alias{c_sigma}
\title{Calculates Sigma (standard deviation) of each ant}
\usage{
c_sigma(param_list, eps)
}
\arguments{
\item{param_list}{a list of the x and y values of the ants of the current generation}

\item{eps}{represents the symmetry to the expected value; has a range of (0,1); if eps = 0.5 the distribution is symmetric}
}
\value{
a list with the standard deviations of the x and y values of each ant
}
\description{
Calculates Sigma (standard deviation) of each ant
}
