% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{calc_abs_diff_to_actual_min}
\alias{calc_abs_diff_to_actual_min}
\title{Return a dataframe with the absolute difference of the x, y and z values of the calculated minimum and the actual minimum}
\usage{
calc_abs_diff_to_actual_min(min_aco_df, min_actual)
}
\arguments{
\item{min_aco_df}{a dataframe with the x,y and z values of the calculated minimum}

\item{min_actual}{a dataframe with the x,y and z values of the actual minimum}
}
\description{
Return a dataframe with the absolute difference of the x, y and z values of the calculated minimum and the actual minimum
}
