% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{calc_gens}
\alias{calc_gens}
\title{Calculates an iteration step,  calculate Values for first generation (use Algorithm-Functions from File "fct_aco.R")}
\usage{
calc_gens(
  datos = "NA",
  cost_f,
  param_list,
  gen_p,
  gen,
  q = 0.2,
  eps = 0.5,
  paralelo = 0
)
}
\arguments{
\item{datos}{the data of interest (this parameter is not necessary for us)}

\item{cost_f}{the objective function}
}
\value{
a new list of the x-, y- and f- value of each ant which represents the new locations of the ants
}
\description{
Calculates an iteration step,  calculate Values for first generation (use Algorithm-Functions from File "fct_aco.R")
}
