% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco_evoper.R
\name{calculate_min}
\alias{calculate_min}
\title{ACO Implemenatation with the package evoper}
\usage{
calculate_min(iterations, lower_bound, upper_bound, test_function)
}
\arguments{
\item{iterations}{number of iterations (= generations)}

\item{lower_bound}{the minimum limit of the range in which we search the minimum (for the x1, x2, f value)}

\item{upper_bound}{the maximum limit of the range in which we search the minimum (for the x1, x2, f value)}

\item{test_function}{the objective function to minimize}
}
\value{
the calculated minimum (x1, x2 and f value)
}
\description{
ACO Implemenatation with the package evoper
}
