% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{cost_function_himmelblau}
\alias{cost_function_himmelblau}
\title{Cost function of Himmelblau Function}
\usage{
cost_function_himmelblau(param_list)
}
\arguments{
\item{param_list}{List of Parameters.}
}
\description{
Cost function of Himmelblau Function
}
