% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{get_color}
\alias{get_color}
\title{Get box colour according to the algorithm result (performance tab)}
\usage{
get_color(value, results)
}
\arguments{
\item{value}{the result of the algorithm to evaluate}

\item{results}{all results of all algorithms in a list}
}
\description{
Get box colour according to the algorithm result (performance tab)
}
