% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{get_test_function}
\alias{get_test_function}
\title{Return a test function acording to the function name}
\usage{
get_test_function(function_name, numb_parameters)
}
\arguments{
\item{function_name}{the name of the test function to returns}

\item{numb_parameters}{the number of parameters that the test function takes (vector vs x, y)}
}
\description{
Return a test function acording to the function name
}
