% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{himmelblau_1}
\alias{himmelblau_1}
\title{Himmelblau function with vector parameter}
\usage{
himmelblau_1(param_list)
}
\arguments{
\item{param_list}{The parameters (x, y) to be used for the test function}
}
\description{
Test function for the application of optimisation methods (vectorized)
}
