% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\docType{data}
\name{minima_himmelblau2}
\alias{minima_himmelblau2}
\title{Actual minima of Himmelblau Function in a data frame}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\usage{
minima_himmelblau2
}
\description{
Actual minima of Himmelblau Function in a data frame
}
\keyword{datasets}
