% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco.R
\name{new_gen}
\alias{new_gen}
\title{Calculate the new generation of ants}
\usage{
new_gen(param_list, desv, prob, param_list_r)
}
\arguments{
\item{param_list}{the old generation of ants}

\item{desv}{the calculated standard deviation of each ants x and y values}

\item{param_list_r}{the range of the x, y and f value in which we search the minimum}
}
\value{
a list of the x and y value of each ant of the generation
}
\description{
Calculate the new generation of ants
}
