% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\docType{data}
\name{opt_himmelblau}
\alias{opt_himmelblau}
\title{Minimize Himmelblau function}
\format{
An object of class \code{list} of length 5.
}
\usage{
opt_himmelblau
}
\description{
Minimize Himmelblau function
}
\details{
If convergence=0 function converges; 4 mimima and a global minimum at x1=-3.77, x2=-3.28, f=0
}
\keyword{datasets}
