% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\docType{data}
\name{opt_rosenbrock}
\alias{opt_rosenbrock}
\title{Minimize Rosenbrock function}
\format{
An object of class \code{list} of length 5.
}
\usage{
opt_rosenbrock
}
\description{
Minimize Rosenbrock function
}
\details{
If convergence=0 function converges; Mimimum at x1=1, x2=1, f=0
}
\keyword{datasets}
