% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco.R
\name{pesos}
\alias{pesos}
\title{Calculate the weight of each ant based on the gaussian distribution}
\usage{
pesos(err, q)
}
\arguments{
\item{err}{the calculated list of errors of the ants}

\item{q}{meta-parameter, with a range of (0,1). If q is close to 0 it will give more relevance to better ants,
on the other hand, if its closer to 1 it will distribute the weight between all the ants more equitatively}
}
\value{
a list with the ants' weights
}
\description{
Calculate the weight of each ant based on the gaussian distribution
}
