% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{prepare_for_plot}
\alias{prepare_for_plot}
\title{Prepares the data with the ants' locations for plotting}
\usage{
prepare_for_plot(hor_number, xyf)
}
\arguments{
\item{hor_number}{Number of ants.}

\item{xyf}{the list of the x-, y- and f- value of each ant which represents the locations of the ants}
}
\description{
Prepares the data with the ants' locations for plotting
}
