% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aco.R
\name{prob_hor}
\alias{prob_hor}
\title{Calculates the probability}
\usage{
prob_hor(peso)
}
\arguments{
\item{peso}{the list of the calculated ants' weights}
}
\value{
a list of the probabilities of choosing each ant as a prototype for the next generation
}
\description{
Calculates the probability
}
