% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{return_3d_plot}
\alias{return_3d_plot}
\title{Generate 3d function for a given objective function}
\usage{
return_3d_plot(
  fu = "rosenbrock",
  minim = -1,
  maxim = 1,
  theta_input = "150",
  phi_input = "20",
  shade_input = 0.3,
  colour = "green"
)
}
\arguments{
\item{fu}{the function to plot for}

\item{minim}{Minimum limit of the axes}

\item{maxim}{Maximum limit of the axes}

\item{theta_input}{Angle defining the viewing direction. Theta gives the degree of the vertical rotation.}

\item{phi_input}{Angle defining the viewing direction. Phi gives the degree of the horizontal rotation.}

\item{colour}{The color(s) of the surface facets._input}

\item{shade_inputValues}{Values of shade close to one yield shading similar to a point light source model and values 
close to zero
   produce no shading. Values in the range 0.5 to 0.75 provide an approximation to daylight illumination.}
}
\description{
Generate 3d function for a given objective function
}
