% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{rosenbrock_1}
\alias{rosenbrock_1}
\title{Rosenbrock function with vector parameter}
\usage{
rosenbrock_1(param_list)
}
\arguments{
\item{param_list}{The parameters (x, y) to be used for the test function}
}
\description{
Test function for the application of optimisation methods
}
