% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{rosenbrock_2}
\alias{rosenbrock_2}
\title{Rosenbrock function with extracted parameters}
\usage{
rosenbrock_2(x1, x2)
}
\arguments{
\item{x1, }{x2  The parameters to be used for the test function}
}
\description{
Test function for the application of optimisation methods
}
