% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_controlbar.R
\name{update_controlbar}
\alias{update_controlbar}
\title{Updates the controlbar}
\usage{
update_controlbar(input, output, session)
}
\arguments{
\item{input}{Input from global server.}

\item{output}{Output from global server.}

\item{session}{Session from global server.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Updates the controlbar according to the currently active tab
}
