% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_gens}
\alias{calc_gens}
\title{Values for first generation, use Algorithm-Functions from File "Ameisenalgorythmus_Implementierung_2"}
\usage{
calc_gens(
  datos = "NA",
  cost_f,
  param_list,
  gen_p,
  gen,
  q = 0.2,
  eps = 0.5,
  paralelo = 0
)
}
\description{
Values for first generation, use Algorithm-Functions from File "Ameisenalgorythmus_Implementierung_2"
}
